+++ src/main/java/org/folio/scheduler/integration/kafka/KafkaMessageListener.java
@@ -1,21 +1,23 @@
 package org.folio.scheduler.integration.kafka;
 
+import static java.lang.Boolean.TRUE;
 import static java.util.Collections.singletonList;
+import static org.apache.commons.collections4.CollectionUtils.isNotEmpty;
 import static org.folio.scheduler.utils.OkapiRequestUtils.getStaticPath;
 import static org.folio.spring.integration.XOkapiHeaders.TENANT;
 import static org.folio.spring.integration.XOkapiHeaders.USER_ID;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.Collection;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.apache.kafka.clients.consumer.ConsumerRecord;
+import org.folio.common.utils.SemverUtils;
 import org.folio.scheduler.domain.dto.RoutingEntry;
 import org.folio.scheduler.domain.dto.TimerDescriptor;
 import org.folio.scheduler.integration.kafka.model.ResourceEvent;
-import org.folio.scheduler.integration.kafka.model.ScheduledTimers;
 import org.folio.scheduler.integration.keycloak.SystemUserService;
 import org.folio.scheduler.service.SchedulerTimerService;
 import org.folio.spring.FolioModuleMetadata;
@@ -28,7 +30,6 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 public class KafkaMessageListener {
 
-  private final ObjectMapper objectMapper;
   private final SystemUserService systemUserService;
   private final FolioModuleMetadata folioModuleMetadata;
   private final SchedulerTimerService schedulerTimerService;
@@ -46,15 +47,66 @@ public class KafkaMessageListener {
     concurrency = "#{folioKafkaProperties.listener['scheduled-jobs'].concurrency}")
   public void handleScheduledJobEvent(ConsumerRecord<String, ResourceEvent> consumerRecord) {
     var resourceEvent = consumerRecord.value();
-    var tenantId = resourceEvent.getTenant();
+    var operationType = resourceEvent.getType();
 
-    try (var ignored = new FolioExecutionContextSetter(folioModuleMetadata, prepareContextHeaders(tenantId))) {
-      var scheduledTimers = objectMapper.convertValue(resourceEvent.getNewValue(), ScheduledTimers.class);
-      for (var routingEntry : scheduledTimers.getTimers()) {
-        var routingEntryKey = getRoutingEntryKey(routingEntry);
-        log.info("Processing scheduled job event from kafka [routing entry: '{}']", routingEntryKey);
-        var timerDescriptor = new TimerDescriptor().enabled(true).routingEntry(routingEntry);
+    switch (operationType) {
+      case CREATE:
+        createTimers(resourceEvent);
+        break;
+      case UPDATE:
+        updateTimers(resourceEvent);
+        break;
+      case DELETE:
+        deleteTimers(resourceEvent);
+        break;
+      default:
+        log.warn("Unsupported operation type: consumerRecord = {}", consumerRecord);
+    }
+  }
+
+  private void createTimers(ResourceEvent event) {
+    try (var ignored = new FolioExecutionContextSetter(folioModuleMetadata, prepareContextHeaders(event.getTenant()))) {
+      var moduleName = SemverUtils.getName(event.getNewValue().getModuleId());
+      var timers = event.getNewValue().getTimers();
+      logCreatingTimers(timers);
+      for (var routingEntry : timers) {
+        var timerDescriptor = new TimerDescriptor().enabled(TRUE).moduleName(moduleName).routingEntry(routingEntry);
         schedulerTimerService.create(timerDescriptor);
+      }
+    }
+  }
+
+  private void updateTimers(ResourceEvent resourceEvent) {
+    try (var ignored = new FolioExecutionContextSetter(folioModuleMetadata,
+      prepareContextHeaders(resourceEvent.getTenant()))) {
+      var moduleName = SemverUtils.getName(resourceEvent.getNewValue().getModuleId());
+      var timers = schedulerTimerService.findByModuleName(moduleName);
+      logDeletingTimers(timers);
+      if (isNotEmpty(timers)) {
+        for (var timer : timers) {
+          schedulerTimerService.delete(timer.getId());
+        }
+      }
+
+      var timerToCreate = resourceEvent.getNewValue();
+      logCreatingTimers(timerToCreate.getTimers());
+      for (var routingEntry : timerToCreate.getTimers()) {
+        var timerDescriptor = new TimerDescriptor().enabled(TRUE).moduleName(moduleName).routingEntry(routingEntry);
+        schedulerTimerService.create(timerDescriptor);
+      }
+    }
+  }
+
+  private void deleteTimers(ResourceEvent resourceEvent) {
+    try (var ignored = new FolioExecutionContextSetter(folioModuleMetadata,
+      prepareContextHeaders(resourceEvent.getTenant()))) {
+      var moduleName = SemverUtils.getName(resourceEvent.getOldValue().getModuleId());
+      var timers = schedulerTimerService.findByModuleName(moduleName);
+      logDeletingTimers(timers);
+      if (isNotEmpty(timers)) {
+        for (var timer : timers) {
+          schedulerTimerService.delete(timer.getId());
+        }
       }
     }
   }
@@ -69,5 +121,15 @@ public class KafkaMessageListener {
   private static String getRoutingEntryKey(RoutingEntry routingEntry) {
     var methods = String.join("|", routingEntry.getMethods());
     return methods + " " + getStaticPath(routingEntry);
+  }
+
+  private static void logCreatingTimers(List<RoutingEntry> entries) {
+    var methods = entries.stream().map(KafkaMessageListener::getRoutingEntryKey).toList();
+    log.info("Processing scheduled job event from kafka: timers = {}", methods);
+  }
+
+  private static void logDeletingTimers(List<TimerDescriptor> timers) {
+    var methods = timers.stream().map(t -> getRoutingEntryKey(t.getRoutingEntry())).toList();
+    log.info("Deleting timers: timers = {}", methods);
   }
 }